Fixed call transactional method for save user from same-class methods with incompatible '@Transactional' values
