+++ service/src/main/java/greencity/security/service/GoogleSecurityServiceImpl.java
@@ -2,42 +2,24 @@ package greencity.security.service;
 
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
-import com.google.api.client.http.javanet.NetHttpTransport;
-import com.google.api.client.json.jackson2.JacksonFactory;
 import greencity.client.RestClient;
 import greencity.constant.ErrorMessage;
 import greencity.dto.ubs.UbsProfileCreationDto;
 import greencity.dto.user.UserVO;
-import greencity.entity.Language;
 import greencity.entity.User;
-import greencity.entity.UserAchievement;
-import greencity.entity.UserAction;
-import greencity.enums.EmailNotification;
-import greencity.enums.Role;
 import greencity.enums.UserStatus;
 import greencity.exception.exceptions.UserDeactivatedException;
-import greencity.repository.UserRepo;
 import greencity.security.dto.SuccessSignInDto;
 import greencity.security.jwt.JwtTool;
-import greencity.service.AchievementService;
 import greencity.service.UserService;
-import java.util.UUID;
 
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
 import java.security.GeneralSecurityException;
-import java.time.LocalDateTime;
-import java.util.List;
-
-import static greencity.constant.AppConstant.*;
-import static greencity.security.service.OwnSecurityServiceImpl.getUserAchievements;
-import static greencity.security.service.OwnSecurityServiceImpl.getUserActions;
 
 /**
  * {@inheritDoc}
@@ -49,33 +31,39 @@ public class GoogleSecurityServiceImpl implements GoogleSecurityService {
     private final GoogleIdTokenVerifier googleIdTokenVerifier;
     private final JwtTool jwtTool;
     private final ModelMapper modelMapper;
-    private final AchievementService achievementService;
-    private final UserRepo userRepo;
     private final RestClient restClient;
+    private final GoogleSecurityTransactionalService googleSecurityTransactionalService;
 
     /**
      * Constructor.
      *
-     * @param userService           {@link UserService} - service of {@link User}
-     *                              logic.
-     * @param jwtTool               {@link JwtTool} - tool for jwt logic.
-     * @param googleIdTokenVerifier {@link GoogleIdTokenVerifier} - tool for verify
+     * @param userService                        {@link UserService} - service of
+     *                                           {@link User} logic.
+     * @param jwtTool                            {@link JwtTool} - tool for jwt
+     *                                           logic.
+     * @param googleIdTokenVerifier              {@link GoogleIdTokenVerifier} -
+     *                                           tool for verify.
+     * @param modelMapper                        {@link ModelMapper} - tool for
+     *                                           mapping models.
+     * @param restClient                         {@link RestClient} - tool for
+     *                                           sending requests
+     * @param googleSecurityTransactionalService {@link GoogleSecurityTransactionalService}
+     *                                           - service with transactional
+     *                                           methods
      */
     @Autowired
     public GoogleSecurityServiceImpl(UserService userService,
         JwtTool jwtTool,
         GoogleIdTokenVerifier googleIdTokenVerifier,
         ModelMapper modelMapper,
-        AchievementService achievementService,
-        UserRepo userRepo,
-        RestClient restClient) {
+        RestClient restClient,
+        GoogleSecurityTransactionalService googleSecurityTransactionalService) {
         this.userService = userService;
         this.jwtTool = jwtTool;
         this.googleIdTokenVerifier = googleIdTokenVerifier;
         this.modelMapper = modelMapper;
-        this.achievementService = achievementService;
-        this.userRepo = userRepo;
         this.restClient = restClient;
+        this.googleSecurityTransactionalService = googleSecurityTransactionalService;
     }
 
     /**
@@ -88,12 +76,10 @@ public class GoogleSecurityServiceImpl implements GoogleSecurityService {
             if (googleIdToken != null) {
                 GoogleIdToken.Payload payload = googleIdToken.getPayload();
                 String email = payload.getEmail();
-                String userName = (String) payload.get(USERNAME);
                 UserVO userVO = userService.findByEmail(email);
                 if (userVO == null) {
                     log.info(ErrorMessage.USER_NOT_FOUND_BY_EMAIL + email);
-                    String profilePicture = (String) payload.get(GOOGLE_PICTURE);
-                    User user = saveNewUser(email, userName, profilePicture, language);
+                    User user = googleSecurityTransactionalService.signUp(payload, language);
                     restClient.createUbsProfile(modelMapper.map(user, UbsProfileCreationDto.class));
                     userVO = modelMapper.map(user, UserVO.class);
                     log.info("Google sign-up and sign-in user - {}", userVO.getEmail());
@@ -111,56 +97,6 @@ public class GoogleSecurityServiceImpl implements GoogleSecurityService {
         } catch (GeneralSecurityException | IOException e) {
             throw new IllegalArgumentException(ErrorMessage.BAD_GOOGLE_TOKEN + ". " + e.getMessage());
         }
-    }
-
-    /**
-     * Method that allow save new user during registration via Google with extracted
-     * payload from token.
-     *
-     * @param email          {@link String} - email.
-     * @param userName       {@link String} - name of user.
-     * @param profilePicture {@link String} - profile picture link.
-     * @param language       {@link String} - language.
-     * @return Saved user {@link User}
-     * @author Maksym Golik
-     */
-    @Transactional
-    public User saveNewUser(String email, String userName, String profilePicture,
-        String language) {
-        User user = createNewUser(email, userName, profilePicture, language);
-        user.setUserAchievements(createUserAchievements(user));
-        user.setUserActions(createUserActions(user));
-        user.setUuid(UUID.randomUUID().toString());
-        User savedUser = userRepo.save(user);
-        user.setId(savedUser.getId());
-        return user;
-    }
-
-    private User createNewUser(String email, String userName, String profilePicture,
-        String language) {
-        return User.builder()
-            .email(email)
-            .name(userName)
-            .role(Role.ROLE_USER)
-            .dateOfRegistration(LocalDateTime.now())
-            .lastActivityTime(LocalDateTime.now())
-            .userStatus(UserStatus.ACTIVATED)
-            .emailNotification(EmailNotification.DISABLED)
-            .refreshTokenKey(jwtTool.generateTokenKey())
-            .profilePicturePath(profilePicture)
-            .rating(DEFAULT_RATING)
-            .language(Language.builder()
-                .id(modelMapper.map(language, Long.class))
-                .build())
-            .build();
-    }
-
-    private List<UserAchievement> createUserAchievements(User user) {
-        return getUserAchievements(user, achievementService);
-    }
-
-    private List<UserAction> createUserActions(User user) {
-        return getUserActions(user, achievementService);
     }
 
     private SuccessSignInDto getSuccessSignInDto(UserVO user) {
