+++ service/src/main/java/greencity/security/service/VerifyEmailServiceImpl.java
@@ -13,11 +13,9 @@ import greencity.security.repository.VerifyEmailRepo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
 
 /**
